+++ spec/models/spec_spec.rb
@@ -1,18 +1,18 @@
 require 'spec_helper'
 
 describe Konacha::Spec do
-  describe "#basename" do
-    it "is the basename of the path" do
-      described_class.new("array_spec.js").basename.should == "array_spec"
-      described_class.new("array_spec.coffee").basename.should == "array_spec"
+  describe "#asset_name" do
+    it "is the asset_name of the path" do
+      described_class.new("array_spec.js").asset_name.should == "array_spec"
+      described_class.new("array_spec.coffee").asset_name.should == "array_spec"
     end
 
     it "ignores multiple extensions" do
-      described_class.new("array_spec.js.coffee").basename.should == "array_spec"
+      described_class.new("array_spec.js.coffee").asset_name.should == "array_spec"
     end
 
     it "includes relative path" do
-      described_class.new("subdirectory/array_spec.js").basename.should == "subdirectory/array_spec"
+      described_class.new("subdirectory/array_spec.js").asset_name.should == "subdirectory/array_spec"
     end
   end
 
@@ -31,7 +31,7 @@ describe Konacha::Spec do
   end
 
   describe ".find" do
-    it "returns the Spec with the given basename" do
+    it "returns the Spec with the given asset_name" do
       all = [described_class.new("a_spec.js"),
              described_class.new("b_spec.js")]
       described_class.should_receive(:all) { all }