+++ History.md
@@ -1,6 +1,7 @@
 # master
 
 * Support foo_spec.coffee files (without .js)
+* Rename Konacha::Spec#basename to #asset_name
 
 # 0.9.0